test(river): split LONG_TESTS River pole into tests/river package
Move sync_worker_load_test.go (the 50s concurrent-sync load pole) and
sync_worker_leased_retry_test.go (carries fastRetryPolicy, used by the
load test) into a new tests/river package with its own integration_testdb
TestMain bridge so it clones its own template DB and runs as a concurrent
process. Helpers route through testsupport.
